Update README.md (2 changes)
Added the current working directory `.` to the PATH used by the build/install scripts; otherwise, at least under `bash`, `csh` and `zsh`, the `env` file will not be found in `. env` - this seemed easier than modifying each individual package's `build` and `install` scripts + templates to fully-qualify `. ./env`

Updated dir that needs to be added to `PATH` from `/usr/local/bin` to `/opt/backstep/bin`, which matches PREFIX and I assume is the desired result. Also, the former is already in the default $PATH in NS 3.3.